Добавляет подготовленные выражения
